not quite right last time

+++ scripts/composite-mrc-to-hdf5.py
@@ -98,7 +98,7 @@ fname_fmt = '{0}{1:d}{2}'.format('{0}/{1}{2:0>',digits,'d}.hdf5')
 for k in range(nimg):
 	infile.seek(HEADER_LEN + k*npix)
 	byte_img = infile.read(npix*4)
-	arr_img = numpy.fromstring(byte_img,numpy.double)
+	arr_img = numpy.fromstring(byte_img,numpy.float32).astype(numpy.double)
 
 	arr_img.shape = nrow,ncol
